Move media metric name test to unit tests
